This change introduces history.replaceState on top of our already-functional hash-driven navigation to  clean-up hash-based external URLs as hash change events occur. I've found this behaves very well on properly-supporting browsers like Chrome and Firefox, while falling back to hash-based urls on non-supporting platforms, allowing us to preserve deep links to Ajax-driven pages (at least in JS-enabled environments).
This approach seems to be particularly useful on iOS, where pushState is supported, but often does nothing due to security restrictions that require pushState calls to occur only as the result of a user interaction.

Note that replaceState is only used for externally-available URLs, so dialogs and generated pages like nested listviews still require hash-based links (because the pages are actually located within the page referenced in the URL that precedes the hash.