This change moves the ajax navigation opt-out list (which contains, mainly, browsers that have broken history tracking of hash changes) over to the support.js file, and added a multi-tierd Nokia condition that checks a number of UA conditions to try and pinpoint Nokia + Symbian + Webkit + 7.3-or-older. Big thanks to Bryan and Stephanie at Yiibu for their helpful article on Nokia device detection. That said, we would LOVE to find a non-UA based way to handle this detection. Reluctantly, this seems to do what we need at the moment, while assuming hash history will be improved in any post-7.3 release.
Also included in this commit is a Nokia-specific condition that re-appends existing stylesheets to force them to render, as all pre-7.3 devices we tested would not render the CSS if it was referenced before the JavaScript (as we'd recommend doing). There may be a non-JS way to work around this problem, but for now, this seems to do the trick.